+++ verifiers/utils/error_utils.py
@@ -1,5 +1,3 @@
-from typing import Callable
-
 import verifiers as vf
 
 
@@ -19,36 +17,3 @@ def get_error_chain(
 def get_vf_error_chain(error: BaseException) -> list[vf.Error]:
     """Get an error chain containing only vf errors."""
     return get_error_chain(error, parent_type=vf.Error)
-
-
-class ErrorChain:
-    """Helper class for error chains."""
-
-    def __init__(
-        self,
-        error: BaseException,
-        build_error_chain: Callable[
-            [BaseException], list[BaseException]
-        ] = get_error_chain,
-    ):
-        self.root_error = error
-        self.chain = build_error_chain(error)
-
-    def __hash__(self) -> int:
-        return hash(tuple(type(e).__name__ for e in self.chain))
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, ErrorChain):
-            return NotImplemented
-        return tuple(type(e).__name__ for e in self.chain) == tuple(
-            type(e).__name__ for e in other.chain
-        )
-
-    def __contains__(self, error: BaseException) -> bool:
-        return any(isinstance(error, e) for e in self.chain)
-
-    def __repr__(self) -> str:
-        return " -> ".join([type(e).__name__ for e in self.chain])
-
-    def __str__(self) -> str:
-        return ", caused by ".join([repr(e) for e in self.chain])
