+++ verifiers/utils/logging_utils.py
@@ -8,9 +8,7 @@ from rich.panel import Panel
 from rich.table import Table
 from rich.text import Text
 
-from verifiers.errors import Error
-from verifiers.types import Messages
-from verifiers.utils.error_utils import ErrorChain
+import verifiers as vf
 
 
 def setup_logging(
@@ -43,9 +41,9 @@ def setup_logging(
 
 
 def print_prompt_completions_sample(
-    prompts: list[Messages],
-    completions: list[Messages],
-    errors: list[Error | None],
+    prompts: list[vf.Messages],
+    completions: list[vf.Messages],
+    errors: list[vf.Error | None],
     rewards: list[float],
     step: int,
     num_samples: int = 1,
@@ -103,7 +101,7 @@ def print_prompt_completions_sample(
 
     def _format_error(error: BaseException) -> Text:
         out = Text()
-        out.append(f"error: {ErrorChain(error)}", style="bold red")
+        out.append(f"error: {vf.ErrorChain(error)}", style="bold red")
 
         return out
